# gaggiuino-local-profiler/go/internal/system/openapi.yaml
openapi: 3.0.3
info:
  title: GLP REST API
  version: 2.0.0
  description: >
    REST API for the **Gaggiuino Local Profiler** Home Assistant add-on.
    Provides shot storage, coffee library, maintenance tracking, a barista
    orders system, and live machine data.
  contact:
    url: https://github.com/mxkissnr/gaggiuino-local-profiler
servers:
  - url: http://localhost:8099
    description: Direct port (integration, cards)
  - url: /api/hassio_ingress/gaggiuino_local_profiler
    description: HA Ingress (browser — auth via HA session, no token needed)
security:
  - tokenAuth: []
components:
  securitySchemes:
    tokenAuth:
      type: apiKey
      in: header
      name: X-GLP-Token
      description: >
        64-char hex token, auto-generated on first start, stored in
        `/data/api_token.txt`. Retrieve from `GET /api/status`.
        Requests via HA Ingress bypass this check (HA session is sufficient).
        `GET /api/status` is always public.
    tokenQueryAuth:
      type: apiKey
      in: query
      name: token
      description: >
        Same token as `tokenAuth`, accepted as a `?token=` query parameter
        instead of the `X-GLP-Token` header. Only honored on `GET
        /api/events`, since a browser-native EventSource connection can't
        set custom headers — every other route requires the header.
  schemas:
    Ok:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean, example: true }
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, example: "not found" }
...
    MaintenanceTask:
      type: object
      description: >
        One maintenance task's computed state. The base keys are always sent;
        threshold_g/grinderName/disabled are grinder_* only, label is custom_*
        only, machineSyncedAt is descaling/backflush only, and gramsSince is
        only added for grinder_* tasks that track a gram threshold.
      required: [lastDate, threshold_shots, threshold_days, daysSince, shotsSince, pct, status]
      properties:
        lastDate:        { type: string, nullable: true, description: "ISO timestamp of the last completion (a plain YYYY-MM-DD after a restore/import), null when never done" }
        threshold_shots: { type: integer, minimum: 1, maximum: 10000, nullable: true }
        threshold_days:  { type: integer, minimum: 1, maximum: 3650, nullable: true }
        threshold_g:     { type: integer, minimum: 1, maximum: 100000, nullable: true, description: "grinder_* only: gram threshold backing gramsSince" }
        grinderName:     { type: string, description: "grinder_* only: the grinder's display name" }
        label:           { type: string, description: "custom_* only: the user-defined label" }
        machineSyncedAt: { type: string, nullable: true, description: "descaling/backflush only: when the task was last auto-synced from the machine" }
        disabled:        { type: boolean, description: "grinder_* only: task hidden from the maintenance view when true" }
        daysSince:       { type: integer, nullable: true, description: "Days since last maintenance, null when never done" }
        shotsSince:      { type: integer, description: Shots since last maintenance }
        gramsSince:      { type: integer, description: "grinder_* only: grams dosed since last maintenance" }
        pct:             { type: number, minimum: 0, maximum: 1, description: "Progress toward threshold (0–1, capped at 1)" }
        status:          { type: string, enum: [never, ok, soon, due] }
    MaintenanceStats:
      type: object
      description: >
        Keys are task names: `descaling`, `backflush`, `grouphead`,
        `gaskets`, `waterfilter`, `grinder_{id}`.
      additionalProperties:
        $ref: '#/components/schemas/MaintenanceTask'
    MaintenanceAllMachinesStats:
      type: object
      description: >
        Shape returned by GET /api/maintenance?machineId=all (#392) — distinct
        from MaintenanceStats, not a map keyed by machine.
      properties:
        all: { type: boolean, example: true }
        machines:
          type: array
          items:
            type: object
            properties:
              machineId:   { type: integer }
              machineName: { type: string }
              tasks:
                type: object
                description: Per-machine-scoped tasks only (descaling/backflush/grouphead/gaskets)
                additionalProperties: { $ref: '#/components/schemas/MaintenanceTask' }
        global:
          type: object
          description: Shared-equipment tasks (waterfilter/grinder_*), identical regardless of machine
          additionalProperties: { $ref: '#/components/schemas/MaintenanceTask' }
    MaintenanceLogEntry:
      type: object
      required: [id, ts, date, task]
      properties:
        id:        { type: integer }
        ts:        { type: integer, description: Unix ms }
        date:      { type: string, format: date }
        task:      { type: string }
        machine:   { type: string, description: Machine hostname at log time, cosmetic }
        notes:     { type: string, maxLength: 500 }
        shotCount: { type: integer }
        machineId: { type: integer }
        grinderName: { type: string, description: "Resolved display name for a grinder_* task, added on read" }
        label:       { type: string, description: "Resolved display label for a custom_* task, added on read" }

    # ── Orders ────────────────────────────────────────────────────────────
...
paths:
...
  /api/maintenance:
    get:
      tags: [Maintenance]
      summary: Maintenance state with computed stats, for one machine or all machines
      description: >
        Static tasks: `descaling`, `backflush`, `grouphead`, `gaskets`, `waterfilter`.
        Dynamic grinder tasks use the key `grinder_{id}`. `?machineId=all`
        (#392) returns the distinct MaintenanceAllMachinesStats shape instead
        (per-machine-scoped tasks grouped under `machines[]`, shared-equipment
        tasks computed once under `global`). Omitted machineId defaults to 1
        (the default machine) for backward compatibility with older clients.
      parameters:
        - { name: machineId, in: query, schema: { oneOf: [{ type: integer }, { type: string, enum: [all] }] } }
      responses:
        "200":
          description: Maintenance stats
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/MaintenanceStats'
                  - $ref: '#/components/schemas/MaintenanceAllMachinesStats'
  /api/maintenance/{task}/done:
    post:
      tags: [Maintenance]
      summary: Mark a maintenance task as done today
      parameters:
        - name: task
          in: path
          required: true
          schema: { type: string }
          description: Task name (e.g. "descaling", "grinder_1716000000000")
        - { name: machineId, in: query, schema: { type: integer }, description: "Defaults to 1 (the default machine); 'all' is not accepted here, unlike GET /api/maintenance" }
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { notes: { type: string } } }
      responses:
        "200": { description: Updated stats, content: { application/json: { schema: { $ref: '#/components/schemas/MaintenanceStats' } } } }
        "404": { description: Unknown task, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/maintenance/{task}/threshold:
    post:
      tags: [Maintenance]
      summary: Update reminder thresholds for a maintenance task
      parameters:
        - name: task
          in: path
          required: true
          schema: { type: string }
        - { name: machineId, in: query, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                threshold_shots: { type: integer, minimum: 1, maximum: 10000, nullable: true }
                threshold_days:  { type: integer, minimum: 1, maximum: 3650, nullable: true, description: "Static/grinder_* tasks are additionally capped at 365 server-side; custom_* tasks accept the full range." }
                label: { type: string, maxLength: 100, description: "custom_* tasks only — renames the task's display label." }
      responses:
        "200": { description: Updated stats, content: { application/json: { schema: { $ref: '#/components/schemas/MaintenanceStats' } } } }
        "404": { description: Unknown task, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/maintenance/custom:
    post:
      tags: [Maintenance]
      summary: Create a user-defined maintenance task
      parameters:
        - { name: machineId, in: query, schema: { type: integer }, description: "Defaults to 1" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label]
              properties:
                label:           { type: string, maxLength: 100 }
                threshold_shots: { type: integer, minimum: 1, maximum: 10000, nullable: true }
                threshold_days:  { type: integer, minimum: 1, maximum: 3650, nullable: true }
      responses:
        "200": { description: Updated stats, content: { application/json: { schema: { $ref: '#/components/schemas/MaintenanceStats' } } } }
        "400": { description: Missing/empty label, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        "409": { description: A task with the slugified key already exists, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/maintenance/custom/{key}:
    delete:
      tags: [Maintenance]
      summary: Delete a user-defined maintenance task
      parameters:
        - { name: key, in: path, required: true, schema: { type: string } }
        - { name: machineId, in: query, schema: { type: integer }, description: "Defaults to 1" }
      responses:
        "200": { description: Updated stats, content: { application/json: { schema: { $ref: '#/components/schemas/MaintenanceStats' } } } }
        "400": { description: key is not a valid custom_* task key, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/maintenance/log:
    get:
      tags: [Maintenance]
      summary: Maintenance log entries (#393)
      parameters:
        - { name: machineId, in: query, schema: { type: integer }, description: "Omitted or 'all' returns every machine's entries" }
      responses:
        "200": { description: Array of log entries, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/MaintenanceLogEntry' } } } } }
    post:
      tags: [Maintenance]
      summary: Add a maintenance log entry (without necessarily marking the task done)
      parameters:
        - { name: machineId, in: query, schema: { type: integer }, description: "Defaults to 1" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [task]
              properties:
                task:  { type: string }
                date:  { type: string, format: date }
                notes: { type: string, maxLength: 500 }
      responses:
        "200": { description: Created entry, content: { application/json: { schema: { $ref: '#/components/schemas/MaintenanceLogEntry' } } } }
        "400": { description: Invalid task or date }
  /api/maintenance/log/{id}:
    delete:
      tags: [Maintenance]
      summary: Delete a maintenance log entry
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Success, content: { application/json: { schema: { $ref: '#/components/schemas/Ok' } } } }
        "400": { description: Invalid id }
        "404": { description: Not found }

  # ── Orders ───────────────────────────────────────────────────────────────
...